Fix various warnings and upgrade catch2 to 3.8.0 version

+++ tutorial/imgproc/brightness/tutorial-compare-auto-gamma.cpp
@@ -122,7 +122,7 @@ int main(int argc, const char **argv)
     }
     else if (std::string(argv[i]) == "--gaussian-std" && i + 1 < argc) {
       ++i;
-      gaussianStdev = std::atof(argv[i]);
+      gaussianStdev = static_cast<float>(std::atof(argv[i]));
     }
     else if (std::string(argv[i]) == "--aperture-size" && i + 1 < argc) {
       ++i;
